refactor: modularize service classes
Broke down larger service classes into smaller, more focused modules. This improves code organization, makes it easier to test, and allows for better reuse of specific functionalities.
Ex
