+++ src/services/application/interfaces/IApplicationsService.ts
@@ -1,4 +1,4 @@
-import type { SearchResult } from "../search/interfaces/SearchResult";
+import type { SearchResult } from "../../search/interfaces/SearchResult";
 
 /**
  * Interface for managing and interacting with system applications
