+++ src/services/extension/commandService.ts
@@ -1,7 +1,7 @@
 import type { CommandHandler, CommandMatch, CommandMatcher } from "asyar-api";
 import type { ICommandService } from "asyar-api";
 import { writable, get } from "svelte/store";
-import { logService } from "../logService";
+import { logService } from "../log/logService";
 import Fuse from "fuse.js";
 
 interface RegisteredCommand {
