+++ src/services/index.ts
@@ -1,8 +1,8 @@
 // Export all services
-export { logService } from "./logService";
+export { logService } from "./log/logService";
 export { default as extensionManager } from "./extension/extensionManager";
-export { settingsService } from "./settingsService";
-export { ClipboardHistoryService as clipboardService } from "./clipboardHistoryService";
-export { actionService } from "./actionService";
+export { settingsService } from "./settings/settingsService";
+export { ClipboardHistoryService as clipboardService } from "./clipboard/clipboardHistoryService";
+export { actionService } from "./action/actionService";
 export { default as commandService } from "./extension/commandService";
 // Add other services as needed
