+++ src/services/window/interfaces/IWindowService.ts
