+++ src/services/window/windowService.ts
