--- src/types/index.ts
@@ -1,1 +1,0 @@
-export type { AppSettings } from "./AppSettingsType";
